finish renaming spanlink, style dropdown
